// src/app/modules/grade-management/models/grade-promotion-info.model.ts
export class GradePromotionInfo {
    saupjangcd: string = '';
    asscd: string = '';
    gradecd: string = '';
    startymd: string = '';
    endymd: string = '';
    cycle: string = '';
    status: boolean = false;
    accumulate_point: boolean = false;
    ratio_accumulate_point: string = '';
    discount: boolean = false;
    ratio_discount: string = '';
    note: string = '';
    insymd: string = '';
    inssabun: string = '';
    lastymd: string = '';
    lastsabun: string = '';
}